+++ server/index.js
@@ -55,6 +55,9 @@ io.on('connection', (socket) => {
 
     // Confirm the player has been created.
     io.to(socket.id).emit('playerCreated', name);
+
+    // For the current player, spawn all previously existing players in game.
+    io.to(socket.it).emit('spawnAllPreviousPlayers', dataStore.players);
   });
 
   socket.on('disconnect', () => {